Added UI to level 1
UI is hidden in editor view. UI loaded fine in the level, though the pause function does not seem to be working as intended, which I will get to working on soon. I also acknowledge that the pause menu is open on start, which I can fix. If you see it, press the escape key to get rid of it. When I fix the bugs, I will get to working on functionality for the gameplay UI as well.