Stumped + Quick Keys
I am sorry to say, but this issue has me stumped, and it doesn't seem like there are other scripts that are causing the buttons to not render the mouse clicks. I think I will need help trouble shooting this problem. While I am still trying to figure this out, I added a script that allows for quick keys. The inputs do not have functionality yet, but they are to be a keyboard button alternative for building selection, when placing towers.