WinLose Complete
There are now set conditions for winning and losing a level. Note that I do acknowledge that the MainMenu scene gets called around 100 times (which will be fixed) but for the most part, no errors were present related to winning or losing. I might change the script later on so that the player is sent directly to the level select screen. I had been focusing on getting the WinLose script to work, and now I can focus back I on UI gameplay revamp.