Gameplay UI revamped
It is not fully complete however, as most of the buttons need to be assigned with their affiliated towers and traps. I tried to find the array/list of tower objects so that I could modify it myself but I could not find it. I also modified the scripting for tower placement so that it could be called as a method rather than already being in Update(). I acknowledge that there is a bug with tower placement from the drop menus (details will be in the Issues section on Github). I will also add a drop menu for the walls.