Changes to Main Menu and Drop wall
I was preoccupied with another assignment, so I still need to fix that bug, but I at least wanted to add some changes to some UI elements, like the images provided for the Main Menu. I also changed the gameplay UI so that the button icons are towers as opposed to numbers. I changed one of the drop down menus so that it is opted for walls instead of the AOE tower. Also, I added the rest of the levels to the build settings so that the player can access them from the Main Menu. Will do more changes this weekend.