Game progress bug
I think I will need help with this one. I tried to see what is going on with the GameProgress and why it isn't recording level wins, and I can't seem to figure out why that is. The first start variable seems to work fine, and the script isn't exclusive to any scene, and the getScene() method seems to work fine but yet, it is not working for level wins. For some reason, the WinLose script never attempts to change the variables in GameProgress, so the only idea I have for why this is is that it might be related to the boolean variables are static, or the code for getting buttons active doesn't work due to the level buttons being in a seperate scene from the level scenes.