Bug fix, End screen and tutorial WIP
Fixed a bug that placed a bugged manatower when selecting a shoot tower or spike trap from the drop down menus.
I also made a functioning end screen for the game. Beat level 5 or manually change the boolean variable from GameProgress to see it. This was a quick addition, so let me know if this should be developed more.
I also started working on the tutorial or 'how to play' section in the Main Menu. So far, only the first page was made, and I will have to make the rest and make the appropriate screenshots.