Tower fix and potential fix for end screen
Fixed the shoot drop down button where it had placed a mana tower instead of a shoot tower. I also made a potential fix for the endscreen not showing up after beating level 5, though I have not yet tested through level 5 after making this change. If it is not showing up, I'll have to mess around with it some more.