UI change, more comments, and removed unused elements
Replaced the UI sprites with the new ones. I added more documentation to all of the scripts I worked on.
I also went ahead and removed elements like the top bar and structure UI that are not being used at this stage of development.
This should be my last commit to the game, but if anything goes wrong, let me know.